// packages/typedoc-docusaurus-theme/src/types/options.ts
/*
 * THIS FILE IS AUTO GENERATED FROM THE OPTIONS CONFIG. DO NOT EDIT DIRECTLY
 */

/**
 * Describes the options declared by the plugin.
 */
export interface PluginOptions {
  /**
   * The value of the Docusaurus `docs.path` config option.
   */
  docsPath?: string;

  /**
   * The value of the Docusaurus `numberPrefixParser` option.
   */
  numberPrefixParser?: boolean;

  /**
   * Configures the autogenerated sidebar.
   */
  sidebar?: Sidebar;
}

export interface Sidebar {
  autoConfiguration: boolean;
  pretty: boolean;
  typescript: boolean;
  deprecatedItemClassName: string;
}