// packages/typedoc-docusaurus-theme/src/types/sidebar.ts
export type SidebarsConfig = {
  [sidebarId: string]: SidebarConfig | undefined;
};

export type SidebarConfig =
  | SidebarItemConfig[]
  | {
      categories?: SidebarCategoryConfig[];
      items?: SidebarItemConfig[];
      link?: SidebarLinkConfig;
    };

export type SidebarItemConfig =
  | SidebarDocConfig
  | SidebarLinkConfig
  | SidebarCategoryConfig
  | SidebarAutogeneratedConfig;

export type SidebarDocConfig = {
  type: 'doc';
  id: string;
  label?: string;
};

export type SidebarLinkConfig = {
  type: 'link';
  label: string;
  href: string;
};

export type SidebarCategoryConfig = {
  type: 'category';
  label: string;
  items: SidebarItemConfig[];
  collapsed?: boolean;
  collapsible?: boolean;
  link?: SidebarLinkConfig | SidebarDocConfig | SidebarAutogeneratedConfig;
};

export type SidebarAutogeneratedConfig = {
  type: 'autogenerated';
  dirName: string;
};